+++ app/src/main/java/com/yongjincompany/hackerthonandroid/features/calendar/utils/CalendarBindingAdapter.kt
@@ -1,4 +1,4 @@
-package com.yongjincompany.hackerthonandroid.features.home.view.utils
+package com.yongjincompany.hackerthonandroid.features.calendar.utils
 
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
@@ -8,7 +8,7 @@ import com.yongjincompany.hackerthonandroid.common.getDayOfWeek
 import com.yongjincompany.hackerthonandroid.common.getParentActivity
 import java.time.LocalDate
 
-@BindingAdapter("home_date")
+@BindingAdapter("calendar_date")
 fun setMutableDate(view: TextView, date: MutableLiveData<String>) {
     val parentActivity: AppCompatActivity = view.getParentActivity() ?: return
