+++ app/src/main/java/com/yongjincompany/hackerthonandroid/features/calendar/vm/CalendarViewModel.kt
@@ -1,10 +1,10 @@
-package com.yongjincompany.hackerthonandroid.features.home.vm
+package com.yongjincompany.hackerthonandroid.features.calendar.vm
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import java.time.LocalDate
 
-class HomeViewModel : ViewModel() {
+class CalendarViewModel : ViewModel() {
 
     val currentDate = MutableLiveData(LocalDate.now().toString())
     val info = MutableLiveData("입력된 정보가 없습니다.")
