+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,10 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/nav_graph"
-    app:startDestination="@id/homeFragment">
-    <fragment
-        android:id="@+id/homeFragment"
-        android:name="com.yongjincompany.hackerthonandroid.features.home.view.fragment.HomeFragment"
-        android:label="HomeFragment" />
+    android:id="@+id/nav_graph">
+
 </navigation>
